+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 <<<<<<< HEAD
+### 2016-11-02
+* Flight Director fixes for creating DB and Maratyjon reconciliation on start up
+* Sets Flight Director to use localhost for Mesos instead of buggy ELB
+ 
 ### 2016-10-30
 * Logging fixes to for UW1, LVC token support, and Splunk Forwarder OOM issues